#3 list of posts component shows different posts depending on component in which it's in
