+++ src/app/home/home.component.css
@@ -1,0 +1,3 @@
+.home.header {
+    padding-bottom: 1.5em;
+}
