+++ src/app/home/home.component.html
@@ -1,13 +1,13 @@
-<div *ngIf="authService.loggedIn()">
+<div class="home header" *ngIf="authService.loggedIn()">
     <h2>
         Hello, {{ authService.getName() }}.
         <button class="ui positive right floated right labeled icon button" #addPostButton>Add Post<i class="plus icon"></i></button>
     </h2>
 </div>
 
-<app-list-of-posts [componentName]="this.constructor.name" [posts]="posts" [pages]="pages" [currentPage]="currentPage"></app-list-of-posts>
+<app-list-of-posts [componentName]="this.constructor.name"></app-list-of-posts>
 
-<app-error-message *ngIf="isLoaded && !posts.length"
+<app-error-message *ngIf="postService.homeLoaded && !postService.allPosts?.length"
     header="There doesn't seem to be any posts at the moment."
     message="Is the website dead?">
 </app-error-message>
