+++ src/app/home/home.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit, ViewChild, ElementRef, Renderer } from '@angular/core';
 import { AuthService } from '../services/auth.service';
 import { PostService } from '../services/post.service';
-import { ActivatedRoute, Router } from '@angular/router';
 declare var $: any;
 
 @Component({
@@ -11,37 +10,19 @@ declare var $: any;
 })
 export class HomeComponent implements OnInit {
 
-  posts: Array<any> = [];
   pages: Array<number> = []
   currentPage: number;
-  isLoaded: boolean = false;
   @ViewChild('addPostButton') addPostButton: ElementRef;
 
   constructor(
-    private postService: PostService,
     private authService: AuthService,
+    private postService: PostService,
     private renderer: Renderer,
-    private router: Router,
-    private route: ActivatedRoute
   ) {
-    this.route.queryParams.subscribe(params => this.currentPage = params.page);
-    this.currentPage = (!this.currentPage) ? 1 : this.currentPage;
-    this.router.routeReuseStrategy.shouldReuseRoute = () => false;
+    this.postService.homeLoaded = false;
   }
 
   ngOnInit() {
-    this.countPosts();
-    this.getPosts();
-
-    this.postService.postAdded_Observable.subscribe(res => {
-      this.countPosts();
-      this.getPosts();
-    })
-
-    this.postService.postDelete_Observable.subscribe(res => {
-      this.countPosts();
-      this.getPosts();
-    })
   }
 
   ngAfterViewInit() {
@@ -55,37 +36,6 @@ export class HomeComponent implements OnInit {
             .modal('show');
       })
     }
-  }
-
-  getPosts() {
-    if (this.currentPage < 1 || !Number.isInteger(Number(this.currentPage))) {
-      this.currentPage = 1;
-      this.router.navigate([''], { queryParams: { page: this.currentPage } });
-      return;
-    }
-    let pageOffset = (this.currentPage - 1).toString()
-    this.postService.getPosts(pageOffset).subscribe(
-      res => {
-        this.posts = res;
-        this.isLoaded = true;
-        if (!this.posts.length && this.currentPage != 1) {
-          let maxPage = this.pages[this.pages.length - 1];
-          this.currentPage = (this.currentPage > maxPage) ? maxPage : 1;
-          this.router.navigate([''], { queryParams: { page: this.currentPage } });
-        }
-      },
-      err => console.log(err)
-    )
-  }
-
-  countPosts() {
-    this.postService.countPosts().subscribe(
-      res => {
-        let numberOfPages = Math.ceil(res/10);
-        this.pages = Array.from(Array(numberOfPages)).map((x, i) => i + 1);
-      },
-      err => console.log(err)
-    )
   }
 
 }
