+++ src/app/user-profile/user-profile.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { UserService } from '../services/user.service';
-import { Router, ActivatedRoute } from '@angular/router';
-import { PostService } from '../services/post.service';
-import { Post } from '../models/post';
+import { ActivatedRoute } from '@angular/router';
 declare var $: any;
 
 @Component({
@@ -21,34 +19,22 @@ export class UserProfileComponent implements OnInit {
   imageSource: string;
   defaultImageSource: string = 'assets/images/default.png';
 
-  posts: Array<any> = [];
   pages: Array<number> = [];
   currentPage: number;
 
   userDoesNotExist: boolean = false;
-  isLoaded: boolean = false;
   isOwnProfile = false;
 
   constructor(
     private userService: UserService,
-    private postService: PostService,
-    private router: Router,
     private route: ActivatedRoute
   ) {
     this.username = this.route.snapshot.paramMap.get('username');
-    this.route.queryParams.subscribe(params => this.currentPage = params.page);
-    this.currentPage = (!this.currentPage) ? 1 : this.currentPage;
-    this.router.routeReuseStrategy.shouldReuseRoute = () => false;
+    this.userService.profileLoaded = false;
   }
 
   ngOnInit() {
-    this.getUserPostCount();
     this.getUserInfo();
-    
-    this.postService.postDelete_Observable.subscribe(res => {
-      this.getUserPostCount();
-      this.getUserPosts();
-    })
   }
 
   getUserInfo() {
@@ -57,13 +43,12 @@ export class UserProfileComponent implements OnInit {
         this.user = res.user;
         this.imageSource = (this.user['base64']) ? 'data:image/png;base64,' + this.user['base64'] : this.defaultImageSource;
         this.getLoggedInUsername();
-        this.getUserPosts();
       },
       err => {
         console.log(err);
         if (err.status === 404) {
           this.userDoesNotExist = true;
-          this.isLoaded = true;
+          this.userService.profileLoaded = true;
         }
       }
     )
@@ -123,55 +108,6 @@ export class UserProfileComponent implements OnInit {
       },
       err => console.log(err)
     )
-  }
-
-  getUserPosts() {
-    if (this.currentPage < 1 || !Number.isInteger(Number(this.currentPage))) {
-      this.currentPage = 1;
-      this.router.navigate([`u/${this.username}`], { queryParams: { page: this.currentPage } });
-      return;
-    }
-    let pageOffset = (this.currentPage - 1).toString();
-    this.userService.getUserPosts(this.username, pageOffset).subscribe(
-      res => {
-        this.posts = res;
-        this.isLoaded = true;
-        if (!this.posts.length && this.currentPage != 1) {
-          let maxPage = this.pages[this.pages.length - 1];
-          this.currentPage = (this.currentPage > maxPage) ? maxPage : 1;
-          this.router.navigate([`u/${this.username}`], { queryParams: { page: this.currentPage } });
-        }
-      },
-      err => console.log(err)
-    )
-  }
-
-  getUserPostCount() {
-    this.userService.getUserPostCount(this.username).subscribe(
-      res => {
-        let numberOfPages = Math.ceil(res / 10);
-        this.pages = Array.from(Array(numberOfPages)).map((x, i) => i + 1);
-      }
-    )
-  }
-
-  setPostToDelete(post: Post) {
-    this.postService.setPostToDelete(post);
-    $('#deletepost')
-      .modal({
-        transition: 'vertical flip'
-      })
-      .modal('show');
-  }
-
-  setPostToEdit(post: Post) {
-    this.postService.setPostToEdit(post);
-    $('#editpost')
-      .modal({
-        transition: 'slide down',
-        autofocus: false
-      })
-      .modal('show');
   }
 
 }
